pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ross-watchlist-diy"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "ib_insync==0.9.86",
  "requests==2.32.3",
  "python-dotenv==1.0.1",
]

[project.optional-dependencies]
dev = ["ruff", "black", "pytest"]

[project.scripts]
watchlist-build = "watchlist.cli:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
